Enable SKR for both KV and MHSM
Resolves #22750
