+++ sdk/keyvault/test-resources.json
@@ -212,7 +212,6 @@
             "type": "Microsoft.Web/serverfarms",
             "apiVersion": "2020-12-01",
             "name": "[variables('attestationFarm')]",
-            "condition": "[parameters('enableHsm')]",
             "location": "[parameters('location')]",
             "kind": "linux",
             "sku": {
@@ -227,7 +226,6 @@
             "type": "Microsoft.Web/sites",
             "apiVersion": "2020-12-01",
             "name": "[variables('attestationSite')]",
-            "condition": "[parameters('enableHsm')]",
             "dependsOn": [
                 "[resourceId('Microsoft.Web/serverfarms', variables('attestationFarm'))]"
             ],
@@ -281,7 +279,6 @@
         },
         "AZURE_KEYVAULT_ATTESTATION_URL": {
             "type": "string",
-            "condition": "[parameters('enableHsm')]",
             "value": "[format('https://{0}/', reference(variables('attestationSite')).defaultHostName)]"
         }
     }
